+++ src/cli.ts
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 import { Command } from 'commander';
 import { registerConfigCommand } from './commands/config.js';
+import { registerCreateCommand } from './commands/create.js';
 import { registerAttachCommand } from './commands/attach.js';
 import { registerBranchCommand } from './commands/branch.js';
 import { registerCheckoutCommand } from './commands/checkout.js';
@@ -22,6 +23,7 @@ program
   .version('0.1.0');
 
 registerConfigCommand(program);
+registerCreateCommand(program);
 registerAttachCommand(program);
 registerBranchCommand(program);
 registerCheckoutCommand(program);